필요없는 svg파일 제거, 책 상세보기 컴포넌트 분리 작업
